% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_break.R
\name{find_chain_start}
\alias{find_chain_start}
\title{find the start of an infix chain}
\usage{
find_chain_start(doc_lines)
}
\arguments{
\item{doc_lines}{lines of code to examine.}
}
\value{
the index into doc_lines that contains the start of the chain
}
\description{
Working upward from the last line, find the start of the chain.
}
