% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_break.R
\name{get_broken_chain}
\alias{get_broken_chain}
\title{get a broken chain as text}
\usage{
get_broken_chain(doc_lines, doc_cursor_line)
}
\arguments{
\item{doc_lines}{a character vector of R code, one element per line.}

\item{doc_cursor_line}{a number representing the line the cursor is on.}
}
\value{
a character vector of R code representing the broken chain.
}
\description{
This interface is intended for developers who want to hook into the chain
breaking algorithm to create bindings in other text editors.
}
\details{
Given a character vector of R code lines, and the line number of the cursor,
it returns a character vector of R code lines which is the start of the
chained expression the cursor is on, up to the cursor line.

Any assignment with \verb{<-} or \code{=} at the head of the chain is removed.
}
\examples{
get_broken_chain(
    c(
     "species_scatter <- starwars \%>\%",
     "group_by(species, sex) \%>\%",
     "select(height, mass)",
     "    .99s.scatter <- starwars \%>\%",
     "group_by(species, sex) \%>\%",
     "select(height, mass)"
    ),
    3
)
}
