% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfill_base_pipe.R
\name{polyfill_base_pipe}
\alias{polyfill_base_pipe}
\title{Turn "|" operator followed by ">" operator into "|>" operator}
\usage{
polyfill_base_pipe(source_tokens)
}
\description{
Rejig the value columnm of sourcetools::tokenize_string(), such that base
pipe is a first class operator - appears in the \code{value} column as "|>".
}
\details{
This allows all the other infix machinery to handle base pipe as a regular
infix operator.
}
\author{
Miles McBain
}
\keyword{internal}
